added version to homepage, updated login usability
Added version to homepage:
Switched to using a version.php file that can be included where ever its needed
and is the only thing that needs updating for cachebusting and version revs

Updated login usability:
Added "press enter to login" upon focus of the password field. Also reworded
the link to the simplenote create account page.